Refactor trace store meta parse helpers
